\name{OpenFLUID.openProject}
\alias{OpenFLUID.openProject}

\title{Open an existing project from disk}

\description{
blah blah
}

\usage{
  OpenFLUID.openProject(path)
}

\arguments{
\item{path}{Path to an existing project on disk}
}

\seealso{
\code{\link{OpenFLUID.runProject}}
\code{\link{OpenFLUID.runSimulation}}
}

\examples{

}

\keyword{}