\name{OpenFLUID.runProject}
\alias{OpenFLUID.runProject}

\title{Run simulation from an existing project on disk}

\description{
blah blah
}

\usage{
  OpenFLUID.runProject(path)
}

\arguments{
\item{path}{Path to an existing project on disk}
}

\seealso{
\code{\link{OpenFLUID.openProject}}
\code{\link{OpenFLUID.runSimulation}}
}

\examples{

}

\keyword{}