\name{OpenFLUID.runSimulation}
\alias{OpenFLUID.runSimulation}

\title{Run a simulation}

\description{
blah blah
}

\usage{
  OpenFLUID.runSimulation(ofblob)
}

\arguments{
\item{ofblob}{Simulation description to run}
}

\seealso{
\code{\link{OpenFLUID.openProject}}
\code{\link{OpenFLUID.runProject}}
}

\examples{

}

\keyword{}